+++ xbmc/playlists/CMakeLists.txt
@@ -1,5 +1,6 @@
-set(SOURCES PlayListB4S.cpp
-            PlayList.cpp
+set(SOURCES PlayList.cpp
+            PlayListASX.cpp
+            PlayListB4S.cpp
             PlayListFactory.cpp
             PlayListM3U.cpp
             PlayListPLS.cpp
@@ -11,6 +12,7 @@ set(SOURCES PlayListB4S.cpp
             SmartPlaylistFileItemListModifier.cpp)
 
 set(HEADERS PlayList.h
+            PlayListASX.h
             PlayListB4S.h
             PlayListFactory.h
             PlayListM3U.h
