+++ xbmc/playlists/PlayListFactory.cpp
@@ -10,6 +10,7 @@
 
 #include "FileItem.h"
 #include "network/NetworkFileItemClassify.h"
+#include "playlists/PlayListASX.h"
 #include "playlists/PlayListB4S.h"
 #include "playlists/PlayListM3U.h"
 #include "playlists/PlayListPLS.h"
