+++ xbmc/playlists/PlayListPLS.h
@@ -25,14 +25,6 @@ public:
   virtual bool Resize(std::vector<int>::size_type newSize);
 };
 
-class CPlayListASX : public CPlayList
-{
-public:
-  bool LoadData(std::istream &stream) override;
-protected:
-  bool LoadAsxIniInfo(std::istream &stream);
-};
-
 class CPlayListRAM : public CPlayList
 {
 public:
